web/apps/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.2
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
